DI implementation
