+++ app/src/main/java/com/example/android/muicapp/ui/activity/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.example.android.muicapp.activity
+package com.example.android.muicapp.ui.activity
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
@@ -9,13 +9,17 @@ import androidx.navigation.ui.navigateUp
 import androidx.navigation.ui.setupActionBarWithNavController
 import androidx.navigation.ui.setupWithNavController
 import com.example.android.muicapp.R
+import com.example.android.muicapp.di.DaggerAppComponent
+import com.example.android.muicapp.viewmodels.ViewModelProviderFactory
 import kotlinx.android.synthetic.main.activity_main.*
 import kotlinx.android.synthetic.main.app_bar_main.*
+import javax.inject.Inject
 
 class MainActivity : AppCompatActivity() {
 
     lateinit var navController: NavController
     lateinit var appBarConfiguration: AppBarConfiguration
+    val appComponent = DaggerAppComponent.create()
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
