+++ app/src/main/java/com/example/android/muicapp/ui/fragment/about/AboutFragment.kt
@@ -1,0 +1,34 @@
+package com.example.android.muicapp.ui.fragment.about
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.ViewModelProvider
+import com.example.android.muicapp.MyApplication
+import com.example.android.muicapp.R
+import javax.inject.Inject
+
+class AboutFragment : Fragment() {
+
+    @Inject
+    lateinit var providerFactory: ViewModelProvider.Factory
+    lateinit var viewModel: AboutViewModel
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        (activity?.application as MyApplication).appComponent.inject(this)
+        viewModel = ViewModelProvider(this,providerFactory).get(AboutViewModel::class.java)
+        super.onCreate(savedInstanceState)
+    }
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_about, container, false)
+    }
+
+}
