+++ app/src/main/java/com/example/android/muicapp/ui/fragment/about/AboutViewModel.kt
@@ -1,0 +1,7 @@
+package com.example.android.muicapp.ui.fragment.about
+
+import androidx.lifecycle.ViewModel
+import javax.inject.Inject
+
+class AboutViewModel @Inject constructor(): ViewModel() {
+}
