+++ app/src/main/java/com/example/android/muicapp/ui/fragment/samples/SamplesFragment.kt
@@ -1,16 +1,26 @@
-package com.example.android.muicapp.fragment
+package com.example.android.muicapp.ui.fragment.samples
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import androidx.navigation.Navigation
-import androidx.navigation.fragment.findNavController
+import androidx.lifecycle.ViewModelProvider
+import com.example.android.muicapp.MyApplication
 import com.example.android.muicapp.R
-import kotlinx.android.synthetic.main.fragment_samples.*
+import javax.inject.Inject
 
 class SamplesFragment : Fragment() {
+
+    @Inject
+    lateinit var providerFactory: ViewModelProvider.Factory
+    lateinit var viewModel: SamplesViewModel
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        (activity?.applicationContext as MyApplication).appComponent.inject(this)
+        viewModel = ViewModelProvider(this,providerFactory).get(SamplesViewModel::class.java)
+        super.onCreate(savedInstanceState)
+    }
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
