+++ app/src/main/java/com/example/android/muicapp/ui/fragment/samples/SamplesViewModel.kt
@@ -1,0 +1,8 @@
+package com.example.android.muicapp.ui.fragment.samples
+
+import androidx.lifecycle.ViewModel
+import javax.inject.Inject
+
+class SamplesViewModel @Inject constructor(): ViewModel() {
+
+}
